refactor: moving tests around
