+++ crates/arco-core/Cargo.toml
@@ -22,7 +22,3 @@ smallvec = { workspace = true }
 
 [features]
 default = []
-
-[dev-dependencies]
-tracing-subscriber = { workspace = true }
-arco-highs = { workspace = true }
