--- crates/arco-core/examples/README.md
@@ -1,32 +1,0 @@
-# arco-core Examples
-
-This folder contains executable examples demonstrating how to use the `arco-core` crate.
-
-## Available Examples
-
-*To be added:*
-- Basic model building examples
-- Variable and constraint creation
-- Objective function setup
-
-## Running Examples
-
-```bash
-# List available examples
-cargo run --example
-
-# Run a specific example
-cargo run --example <example_name>
-
-# Run with release optimizations
-cargo run --release --example <example_name>
-```
-
-## Notes
-
-The arco-core crate provides the foundation for building optimization models.
-Examples in this folder will demonstrate:
-- Creating models with variables and constraints
-- Setting up objective functions
-- Working with bounds and variable types
-- Model serialization/deserialization
