+++ crates/arco-highs/src/solver.rs
@@ -868,81 +868,4 @@ mod tests {
         let result = outcome.expect("caught panic");
         assert!(result.is_err(), "zero threads should return error");
     }
-
-    /// Helper to build a simple model for testing warm-start
-    fn build_simple_model() -> Model {
-        use arco_core::Constraint;
-
-        let mut model = Model::new();
-
-        let x = model
-            .add_variable(Variable {
-                bounds: Bounds::new(0.0, 10.0),
-                is_integer: false,
-                is_active: true,
-            })
-            .unwrap();
-
-        let y = model
-            .add_variable(Variable {
-                bounds: Bounds::new(0.0, 10.0),
-                is_integer: false,
-                is_active: true,
-            })
-            .unwrap();
-
-        let constraint = model
-            .add_constraint(Constraint {
-                bounds: Bounds::new(0.0, 5.0),
-            })
-            .unwrap();
-
-        model.set_coefficient(x, constraint, 1.0).unwrap();
-        model.set_coefficient(y, constraint, 1.0).unwrap();
-
-        let objective = Objective {
-            sense: Some(Sense::Minimize),
-            terms: vec![(x, 1.0), (y, 1.0)],
-        };
-        model.set_objective(objective).unwrap();
-
-        model
-    }
-
-    #[test]
-    fn test_primal_start_storage() {
-        use arco_expr::VariableId;
-
-        let model = build_simple_model();
-
-        let mut solver = Solver::new(model).unwrap();
-        let hints = vec![(VariableId::new(0), 0.5), (VariableId::new(1), 1.0)];
-        assert!(solver.set_primal_start(&hints).is_ok());
-        assert_eq!(solver.get_primal_start(), Some(hints.as_slice()));
-    }
-
-    #[test]
-    fn test_primal_start_validation() {
-        use arco_expr::VariableId;
-
-        let model = build_simple_model();
-
-        let mut solver = Solver::new(model).unwrap();
-        let invalid_hints = vec![(VariableId::new(9999), 0.5)]; // Non-existent variable
-        assert!(solver.set_primal_start(&invalid_hints).is_err());
-    }
-
-    #[test]
-    fn test_primal_start_clear() {
-        use arco_expr::VariableId;
-
-        let model = build_simple_model();
-
-        let mut solver = Solver::new(model).unwrap();
-        let hints = vec![(VariableId::new(0), 0.5)];
-        solver.set_primal_start(&hints).unwrap();
-        assert!(solver.get_primal_start().is_some());
-        solver.clear_primal_start();
-        assert!(solver.get_primal_start().is_none());
-    }
 }
