+++ crates/arco-solver/Cargo.toml
@@ -16,6 +16,3 @@ workspace = true
 [dependencies]
 arco-core = { workspace = true }
 arco-expr = { workspace = true }
-
-[dev-dependencies]
-tracing-subscriber = { workspace = true }
