--- crates/arco-solver/examples/README.md
@@ -1,32 +1,0 @@
-# arco-solver Examples
-
-This folder contains executable examples demonstrating how to use the `arco-solver` crate.
-
-## Available Examples
-
-*To be added:*
-- Solver abstraction usage examples
-- Custom solver implementations
-- Solution handling examples
-
-## Running Examples
-
-```bash
-# List available examples
-cargo run --example
-
-# Run a specific example
-cargo run --example <example_name>
-
-# Run with release optimizations
-cargo run --release --example <example_name>
-```
-
-## Notes
-
-The arco-solver crate provides shared solver abstractions and traits.
-Examples in this folder will demonstrate:
-- Using the Solver trait
-- Working with Solution objects
-- Solver configuration and options
-- Creating custom solver wrappers
